feat: add number utils
