+++ test/utils/index.ts
@@ -1,4 +1,4 @@
-import { BaseContract, BigNumber } from 'ethers';
+import { BaseContract } from 'ethers';
 import { ethers } from 'hardhat';
 import { Address } from 'hardhat-deploy/types';
 
@@ -41,11 +41,3 @@ export async function setupUser<
     address,
   };
 }
-
-type Input = string | number;
-export const toBN = (value: Input): BigNumber => ethers.BigNumber.from(value);
-export const toGLD = (value: Input): BigNumber =>
-  ethers.utils.parseUnits(
-    typeof value === 'number' ? value.toString() : value,
-    '18'
-  );
